+++ HISTORY.md
@@ -22,3 +22,9 @@ History
 ------------------
 
 * Fixed blank step elements coming in from csv to json conversion
+
+0.0.5 2023-10-19
+* Added waiting steps
+* Added OCR chart data
+* Resolved when no task is present
+* Migrated to a class-based module